make transparent tests generic and move to zcash_client_backend crate
